# crates/wire-schema/openapi.json
{
  "components": {
    "parameters": {
      "ProjectHeader": {
        "description": "Story 0.11 substrate (Phase 3 decision L2). Identifies the target project. Phase 2 single-project deployments accept the header for forward-compatibility: the value must equal the configured project name (case-insensitive after trim) or the reserved sentinel `default`. Mismatching values return 403 `project_not_found`. Absent header is accepted in Phase 2 (with a one-time per-process WARN log) and will become required in Phase 4 multi-project mode.",
        "in": "header",
        "name": "X-OpenGEO-Project",
        "required": false,
        "schema": {
          "type": "string"
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Missing, malformed, or revoked X-OpenGEO-API-Key. The body shape is intentionally opaque so an unauthenticated caller cannot distinguish 'wrong key' from 'auth backend down'."
      }
    },
    "schemas": {
...
      "Error": {
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "error",
          "message"
        ],
        "type": "object"
      },
...
      "GenerateRecommendationsAccepted": {
        "description": "Story 19.6 — 202 response from POST /v1/recommendations/generate. Per the Phase 2 async-write pattern, `status_url` points at the list endpoint where the results are readable.",
        "properties": {
          "generated_count": {
            "minimum": 0,
            "type": "integer"
          },
          "inserted_count": {
            "minimum": 0,
            "type": "integer"
          },
          "status": {
            "type": "string"
          },
          "status_url": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "generated_count",
          "inserted_count",
          "status_url"
        ],
        "type": "object"
      },
...
      "KindAdoption": {
        "type": "object",
        "required": [
          "kind",
          "surfaced",
          "acted",
          "dismissed"
        ],
        "properties": {
          "kind": {
            "type": "string"
          },
          "surfaced": {
            "type": "integer",
            "minimum": 0
          },
          "acted": {
            "type": "integer",
            "minimum": 0
          },
          "dismissed": {
            "type": "integer",
            "minimum": 0
          },
          "adoption_rate": {
            "type": "number",
            "nullable": true
          }
        }
      },
...
      "Recommendation": {
        "description": "Story 19.6 — a stored GEO Recommendation (architecture-phase3-geo-recommendations.md §8 wire shape) plus its DB lifecycle `state`.",
        "properties": {
          "confidence_band": {
            "type": "string"
          },
          "engine_version": {
            "type": "string"
          },
          "generated_at": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "description": "ULID.",
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "payload": {
            "type": "object"
          },
          "project_id": {
            "description": "ULID.",
            "type": "string"
          },
          "reproducibility": {
            "properties": {
              "class": {
                "type": "string"
              },
              "note": {
                "nullable": true,
                "type": "string"
              }
            },
            "required": [
              "class"
            ],
            "type": "object"
          },
          "severity": {
            "type": "string"
          },
          "state": {
            "enum": [
              "generated",
              "surfaced",
              "acknowledged",
              "acted",
              "measured",
              "dismissed",
              "stale"
            ],
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "traceability": {
            "type": "object"
          }
        },
        "required": [
          "id",
          "project_id",
          "kind",
          "severity",
          "confidence_band",
          "state",
          "summary",
          "payload",
          "traceability",
          "reproducibility",
          "tags",
          "generated_at",
          "engine_version"
        ],
        "type": "object"
      },
      "RecommendationIntelligence": {
        "type": "object",
        "required": [
          "by_kind"
        ],
        "properties": {
          "by_kind": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/KindAdoption"
            }
          }
        }
      },
      "RecommendationListResponse": {
        "description": "Story 19.6 — cursor-paginated active recommendations. `next_cursor` is null on the last page.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Recommendation"
            },
            "type": "array"
          },
          "next_cursor": {
            "description": "Opaque cursor for the next page; null when no further pages remain.",
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "type": "object"
      },
...
      "TransitionRecommendationRequest": {
        "description": "Story 19.6 — lifecycle transition (Story 19.4 state machine). Illegal edges return 409.",
        "properties": {
          "evidence_url": {
            "nullable": true,
            "type": "string"
          },
          "note": {
            "nullable": true,
            "type": "string"
          },
          "to": {
            "enum": [
              "surfaced",
              "acknowledged",
              "acted",
              "measured",
              "dismissed",
              "stale"
            ],
            "type": "string"
          }
        },
        "required": [
          "to"
        ],
        "type": "object"
      },
      "TransitionRecommendationResponse": {
        "properties": {
          "recommendation": {
            "$ref": "#/components/schemas/Recommendation"
          },
          "warnings": {
            "items": {
              "type": "object"
            },
            "type": "array"
          }
        },
        "required": [
          "recommendation",
          "warnings"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "in": "header",
        "name": "X-OpenGEO-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Phase 2 v1 surface — read endpoints, prompt run submission, and the SSE event stream. Auth: X-OpenGEO-API-Key header carrying an `ogeo_<32 chars>` key. Story 0.11 (Phase 3 substrate) adds the optional X-OpenGEO-Project header — Phase 2 accepts it for forward compatibility; Phase 4 will require it. Story 15.1 adds three `/v1/setup/*` endpoints (deployment UX substrate).",
    "title": "OpenGEO Public REST API",
    "version": "0.3.1"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/v1/recommendations": {
      "get": {
        "operationId": "listRecommendations",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProjectHeader"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "description": "Opaque page cursor from a previous response's next_cursor.",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationListResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Malformed cursor."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "summary": "Story 19.6 — cursor-paginated active recommendations for the project, newest first."
      }
    },
    "/v1/recommendations/generate": {
      "post": {
        "operationId": "generateRecommendations",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProjectHeader"
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateRecommendationsAccepted"
                }
              }
            },
            "description": "Accepted"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No opengeo.yaml loaded; cannot assemble recommendation inputs."
          }
        },
        "summary": "Story 19.6 — assemble an EngineInput from the project's live prompts/runs/citations, run the in-process engine, and persist the result. Returns 202 + a status_url per the Phase 2 async-write pattern."
      }
    },
    "/v1/recommendations/intelligence": {
      "get": {
        "operationId": "recommendationIntelligence",
        "summary": "Story 19.5+ — per-kind adoption: surfaced / acted / dismissed counts (the what-works-vs-what-doesn't loop). First-party only.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProjectHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationIntelligence"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/v1/recommendations/{id}": {
      "get": {
        "operationId": "getRecommendation",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProjectHeader"
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "description": "Recommendation ULID.",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Recommendation"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Recommendation not found."
          }
        },
        "summary": "Story 19.6 — one recommendation + full traceability. 404 when the row is absent or owned by another project."
      }
    },
    "/v1/recommendations/{id}/state": {
      "patch": {
        "operationId": "transitionRecommendation",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProjectHeader"
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "description": "Recommendation ULID.",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionRecommendationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransitionRecommendationResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unknown target state or malformed id."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Recommendation not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Illegal lifecycle transition."
          }
        },
        "summary": "Story 19.6 — apply a lifecycle transition (Story 19.4 state machine). Illegal transitions return 409."
      }
    },
...
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "servers": [
    {
      "description": "Hosted production",
      "url": "https://api.opengeo.dev"
    },
    {
      "description": "Local Compose stack",
      "url": "http://127.0.0.1:8788"
    }
  ]
}